src/requirements.txt
Pillow
pytesseract
boto3